Let there be a README
And on the first day, he looked upon the stars and noticed that there was a vast space needing to be filled. A space, needing to be described. A space, meant to enrich all other spaces. A space, to define and welcome all newcomers, visitors, returning folks, and those weary travelers simply passing by. He shall call it, README.md